opt/raspi-server/requirements.txt
flask==3.1.2
python-dotenv==1.1.1
firebase-admin==7.1.0
gunicorn==23.0.0
paho-mqtt==2.1.0